fix(elasticloadbalancingv2): logAccessLogs in Base Load Balancer (#6197)
Moving the method logAccessLogs to the Base Load Balancer, so both types
of Load Balancer (Network and Application) can use the method.

closes #3794
