+++ packages/@aws-cdk/aws-elasticloadbalancingv2/lib/shared/base-load-balancer.ts
@@ -1,5 +1,7 @@
 import * as ec2 from '@aws-cdk/aws-ec2';
-import { Construct, IResource, Lazy, Resource } from '@aws-cdk/core';
+import * as iam from '@aws-cdk/aws-iam';
+import * as s3 from '@aws-cdk/aws-s3';
+import { Construct, IResource, Lazy, Resource, Stack, Token } from '@aws-cdk/core';
 import { CfnLoadBalancer } from '../elasticloadbalancingv2.generated';
 import { Attributes, ifUndefined, renderAttributes } from './util';
 
@@ -157,6 +159,34 @@ export abstract class BaseLoadBalancer extends Resource {
   }
 
   /**
+   * Enable access logging for this load balancer.
+   *
+   * A region must be specified on the stack containing the load balancer; you cannot enable logging on
+   * environment-agnostic stacks. See https://docs.aws.amazon.com/cdk/latest/guide/environments.html
+   */
+  public logAccessLogs(bucket: s3.IBucket, prefix?: string) {
+    this.setAttribute('access_logs.s3.enabled', 'true');
+    this.setAttribute('access_logs.s3.bucket', bucket.bucketName.toString());
+    this.setAttribute('access_logs.s3.prefix', prefix);
+
+    const region = Stack.of(this).region;
+    if (Token.isUnresolved(region)) {
+      throw new Error(`Region is required to enable ELBv2 access logging`);
+    }
+
+    const account = ELBV2_ACCOUNTS[region];
+    if (!account) {
+      throw new Error(`Cannot enable access logging; don't know ELBv2 account for region ${region}`);
+    }
+
+    prefix = prefix || '';
+    bucket.grantPut(new iam.AccountPrincipal(account), `${(prefix ? prefix + "/" : "")}AWSLogs/${Stack.of(this).account}/*`);
+
+    // make sure the bucket's policy is created before the ALB (see https://github.com/aws/aws-cdk/issues/1633)
+    this.node.addDependency(bucket);
+  }
+
+  /**
    * Set a non-standard attribute on the load balancer
    *
    * @see https://docs.aws.amazon.com/elasticloadbalancing/latest/application/application-load-balancers.html#load-balancer-attributes
@@ -172,3 +202,30 @@ export abstract class BaseLoadBalancer extends Resource {
     this.setAttribute(key, undefined);
   }
 }
+
+// https://docs.aws.amazon.com/elasticloadbalancing/latest/application/load-balancer-access-logs.html#access-logging-bucket-permissions
+const ELBV2_ACCOUNTS: { [region: string]: string } = {
+  'us-east-1': '127311923021',
+  'us-east-2': '033677994240',
+  'us-west-1': '027434742980',
+  'us-west-2': '797873946194',
+  'ca-central-1': '985666609251',
+  'eu-central-1': '054676820928',
+  'eu-west-1': '156460612806',
+  'eu-west-2': '652711504416',
+  'eu-west-3': '009996457667',
+  'eu-north-1': '897822967062',
+  'ap-east-1': '754344448648',
+  'ap-northeast-1': '582318560864',
+  'ap-northeast-2': '600734575887',
+  'ap-northeast-3': '383597477331',
+  'ap-southeast-1': '114774131450',
+  'ap-southeast-2': '783225319266',
+  'ap-south-1': '718504428378',
+  'me-south-1': '076674570225',
+  'sa-east-1': '507241528517',
+  'us-gov-west-1': '048591011584',
+  'us-gov-east-1': '190560391635',
+  'cn-north-1': '638102146993',
+  'cn-northwest-1': '037604701340',
+};
