+++ packages/@aws-cdk/aws-elasticloadbalancingv2/test/nlb/test.load-balancer.ts
@@ -1,5 +1,6 @@
-import { expect, haveResource } from '@aws-cdk/assert';
+import { expect, haveResource, ResourcePart } from '@aws-cdk/assert';
 import * as ec2 from '@aws-cdk/aws-ec2';
+import * as s3 from '@aws-cdk/aws-s3';
 import * as cdk from '@aws-cdk/core';
 import { Test } from 'nodeunit';
 import * as elbv2 from '../../lib';
@@ -69,6 +70,108 @@ export = {
           Value: "true"
         }
       ]
+    }));
+
+    test.done();
+  },
+
+  'Access logging'(test: Test) {
+    // GIVEN
+    const stack = new cdk.Stack(undefined, undefined, { env: { region: 'us-east-1' } });
+    const vpc = new ec2.Vpc(stack, 'Stack');
+    const bucket = new s3.Bucket(stack, 'AccessLoggingBucket');
+    const lb = new elbv2.NetworkLoadBalancer(stack, 'LB', { vpc });
+
+    // WHEN
+    lb.logAccessLogs(bucket);
+
+    // THEN
+
+    // verify that the LB attributes reference the bucket
+    expect(stack).to(haveResource('AWS::ElasticLoadBalancingV2::LoadBalancer', {
+      LoadBalancerAttributes: [
+        {
+          Key: "access_logs.s3.enabled",
+          Value: "true"
+        },
+        {
+          Key: "access_logs.s3.bucket",
+          Value: { Ref: "AccessLoggingBucketA6D88F29" }
+        }
+      ],
+    }));
+
+    // verify the bucket policy allows the ALB to put objects in the bucket
+    expect(stack).to(haveResource('AWS::S3::BucketPolicy', {
+      PolicyDocument: {
+        Version: '2012-10-17',
+        Statement: [
+          {
+            Action: ["s3:PutObject*", "s3:Abort*"],
+            Effect: 'Allow',
+            Principal: { AWS: { "Fn::Join": ["", ["arn:", { Ref: "AWS::Partition" }, ":iam::127311923021:root"]] } },
+            Resource: {
+              "Fn::Join": ["", [{ "Fn::GetAtt": ["AccessLoggingBucketA6D88F29", "Arn"] }, "/AWSLogs/",
+              { Ref: "AWS::AccountId" }, "/*"]]
+            }
+          }
+        ]
+      }
+    }));
+
+    // verify the ALB depends on the bucket *and* the bucket policy
+    expect(stack).to(haveResource('AWS::ElasticLoadBalancingV2::LoadBalancer', {
+      DependsOn: ['AccessLoggingBucketPolicy700D7CC6', 'AccessLoggingBucketA6D88F29']
+    }, ResourcePart.CompleteDefinition));
+
+    test.done();
+  },
+
+  'access logging with prefix'(test: Test) {
+    // GIVEN
+    const stack = new cdk.Stack(undefined, undefined, { env: { region: 'us-east-1' } });
+    const vpc = new ec2.Vpc(stack, 'Stack');
+    const bucket = new s3.Bucket(stack, 'AccessLoggingBucket');
+    const lb = new elbv2.NetworkLoadBalancer(stack, 'LB', { vpc });
+
+    // WHEN
+    lb.logAccessLogs(bucket, 'prefix-of-access-logs');
+
+    // THEN
+    // verify that the LB attributes reference the bucket
+    expect(stack).to(haveResource('AWS::ElasticLoadBalancingV2::LoadBalancer', {
+      LoadBalancerAttributes: [
+        {
+          Key: "access_logs.s3.enabled",
+          Value: "true"
+        },
+        {
+          Key: "access_logs.s3.bucket",
+          Value: { Ref: "AccessLoggingBucketA6D88F29" }
+        },
+        {
+          Key: "access_logs.s3.prefix",
+          Value: "prefix-of-access-logs"
+        }
+      ],
+    }));
+
+    // verify the bucket policy allows the ALB to put objects in the bucket
+    expect(stack).to(haveResource('AWS::S3::BucketPolicy', {
+      PolicyDocument: {
+        Version: '2012-10-17',
+        Statement: [
+          {
+            Action: ["s3:PutObject*", "s3:Abort*"],
+            Effect: 'Allow',
+            Principal: { AWS: { "Fn::Join": ["", ["arn:", { Ref: "AWS::Partition" }, ":iam::127311923021:root"]] } },
+            Resource: {
+              "Fn::Join": ["", [{ "Fn::GetAtt": ["AccessLoggingBucketA6D88F29", "Arn"] }, "/prefix-of-access-logs/AWSLogs/",
+              { Ref: "AWS::AccountId" }, "/*"]]
+            }
+          }
+        ]
+      }
     }));
 
     test.done();
